third_party/WebKit/LayoutTests/fast/dom/shadow/shadowroot-clonenode.html

<!DOCTYPE>
<html>
<script src="../../js/resources/js-test-pre.js"></script>

<pre id="console"></pre>

<script>
description('Calling ShadowRoot.cloneNode() should throw a DataCloneError exception.');

var host = document.createElement('div');
var shadowRoot = host.createShadowRoot();
shouldThrow('shadowRoot.cloneNode()', '"DataCloneError: An object could not be cloned."');
</script>
<script src="../../js/resources/js-test-post.js"></script>
</html>
